+++ backend/airport/src/main/java/com/thy/airport/repository/CityRepository.java
@@ -5,6 +5,6 @@ import com.thy.airport.entity.City;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface CityRepository extends CrudRepository<City,Long> {
+public interface CityRepository extends CrudRepository<City,Integer> {
 
 }
